added tilemap to the scene so it's accesible everywhere

+++ Minigin/BoundingBoxComponent.h
@@ -12,7 +12,7 @@ namespace comps
 	class BoundingBoxComponent final : public comps::BaseComponent
 	{
 	public:
-		BoundingBoxComponent(std::list < std::shared_ptr<rectangle_>> collisionList, std::shared_ptr<PhysicsComponent> physicsComp,int width,int height);
+		BoundingBoxComponent(std::list < std::shared_ptr<rectangle_>> collisionList, std::shared_ptr<PhysicsComponent> physicsComp,float width,float height);
 		rectangle_ GetBoundingBox(const float elapsedSecs, bool xonly) const;
 		void SetAlignment(HAlign hAlign, VAlign vAlign);
 
@@ -25,8 +25,8 @@ namespace comps
 		std::list < std::shared_ptr<rectangle_>> m_CollisionList;
 		std::shared_ptr<PhysicsComponent> m_pPhysicsComp;
 		std::pair< HAlign, VAlign> m_Alignment;
-		int m_Width;
-		int m_Height;
+		float m_Width;
+		float m_Height;
 
 		bool IsRectangleOverlapping(float elapsedSecs, bool xonly);
